Modify PlaceGamePiece for integration with FieldPositionManager
